:hammer: Remove the platform-specific bindings from SignalDao
